requirements.txt
praw
nltk
flask
matplotlib
mysql-connector-python
python-dotenv
flask-cors
flask-socketio
eventlet
redis